Added status service for colors
