+++ lib/workorder-detail/workorder-detail-controller.js
@@ -7,8 +7,8 @@ var CONSTANTS = require('../constants');
  * @param WORKORDER_CONFIG
  * @constructor
  */
-var statusIconColor = require('../workorder-status/workorder-status-directive.js');
-function WorkorderDetailController(mediator, WORKORDER_CONFIG) {
+
+function WorkorderDetailController(mediator, WORKORDER_CONFIG, workorderStatusService) {
   var self = this;
 
   self.adminMode = WORKORDER_CONFIG.adminMode;
@@ -25,9 +25,9 @@ function WorkorderDetailController(mediator, WORKORDER_CONFIG) {
   };
 
   self.getColorIcon = function(status) {
-    return statusIconColor.getStatusIconColor(status).statusColor;
+    return workorderStatusService.getStatusIconColor(status).statusColor;
   };
 }
 
 
-angular.module(CONSTANTS.WORKORDER_DIRECTIVE).controller('WorkorderDetailController', ["mediator", 'WORKORDER_CONFIG', WorkorderDetailController]);
+angular.module(CONSTANTS.WORKORDER_DIRECTIVE).controller('WorkorderDetailController', ["mediator", 'WORKORDER_CONFIG', 'workorderStatusService', WorkorderDetailController]);
