+++ lib/workorder-list/workorder-list-controller.js
@@ -10,9 +10,9 @@ var CONSTANTS = require('../constants');
  * @constructor
  */
 
-var statusIconColor = require('../workorder-status/workorder-status-directive.js');
+//var statusIconColor = require('../workorder-status/workorder-status-directive.js');
 
-function WorkorderListController(mediator, workorderMediatorService, $q) {
+function WorkorderListController(mediator, workorderMediatorService, $q, workorderStatusService) {
   var self = this;
   var _workorders = [];
 
@@ -45,11 +45,11 @@ function WorkorderListController(mediator, workorderMediatorService, $q) {
   self.getColorIcon = function(workorder) {
     var result = this.resultMap[workorder.id];
     if (!result) {
-      return statusIconColor.getStatusIconColor('').statusColor;
+      return workorderStatusService.getStatusIconColor('').statusColor;
     } else {
-      return statusIconColor.getStatusIconColor(this.resultMap[workorder.id].status).statusColor;
+      return workorderStatusService.getStatusIconColor(this.resultMap[workorder.id].status).statusColor;
     }
   };
 }
 
-angular.module(CONSTANTS.WORKORDER_DIRECTIVE).controller('WorkorderListController', ['mediator', 'workorderMediatorService', '$q', WorkorderListController]);
+angular.module(CONSTANTS.WORKORDER_DIRECTIVE).controller('WorkorderListController', ['mediator', 'workorderMediatorService', '$q', 'workorderStatusService', WorkorderListController]);
