+++ lib/workorder-status/index.js
@@ -1,1 +1,2 @@
 require('./workorder-status-directive');
+require('./workorder-status-service');
